+++ app/src/main/java/com/example/ancientdiggers/data/model/terreno/terrenos/excavables/TerrenoMedioExcavar.kt
@@ -7,10 +7,9 @@ import com.example.ancientdiggers.data.model.terreno.terrenos.TerrenoExcavado
 class TerrenoMedioExcavar (
     nombre: String,
     desbloqueado: Boolean = true,
-    coste: Double
+    coste: Int
 ) : TerrenoExcavable(nombre, desbloqueado, coste){
 
-    // En TerrenoExcavable, la imagen se asigna según el estado
     override var imagen: Int = R.drawable.terreno_medio_excavado
 
     override fun excavar(): TerrenoExcavado {
